src/requirements.txt
# Core dependencies
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
httpx>=0.28
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Vector store dependencies
qdrant-client>=1.7.0
chromadb>=0.4.0
weaviate-client>=4.0.0
numpy>=1.24.0

# Additional dependencies found in imports
# PDF processing dependencies
pdfplumber>=0.9.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
camelot-py>=0.10.0
pandas>=2.0.0
pillow>=10.0.0
langdetect>=1.0.9
opencv-python>=4.8.0

# Markdown and HTML processing
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2020.1.16

# Document processing
python-docx>=1.1.0
openpyxl>=3.1.0
pymdown-extensions>=10.5

# Git processing  
GitPython>=3.1.0

# Data processing
pyyaml>=6.0
python-dateutil>=2.8.0

# Text processing utilities
chardet>=5.2.0
textstat>=0.7.0

# NLP and Knowledge Graph dependencies
spacy>=3.7.0
nltk>=3.8.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
networkx>=3.0
transformers>=4.30.0
torch>=2.0.0

# These will be added as modules are implemented

